backend/requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
asgiref==3.11.1
asttokens==3.0.0
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
bcrypt==5.0.0
beautifulsoup4==4.12.3
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.6
click==8.1.7
colorama==0.4.6
comtypes==1.4.9
constantly==23.10.4
contourpy==1.3.2
cryptography==46.0.6
cycler==0.12.1
daphne==4.2.1
decorator==5.2.1
distro==1.9.0
Django==5.2.12
django-cors-headers==4.9.0
django-extensions==3.2.3
django-filter==24.1
django-ratelimit==4.1.0
django-tailwind==4.4.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
docopt==0.6.2
docstring_parser==0.17.0
earthengine-api==1.7.4
executing==2.2.0
filelock==3.21.2
filetype==1.2.0
Flask==3.1.2
folium==0.19.6
fonttools==4.58.0
fsspec==2026.2.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-genai==1.63.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
gTTS==2.5.3
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
# huggingface_hub==1.4.1 (Conflict)
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.0
Js2Py==0.74
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==1.2.10
langchain-anthropic==1.3.3
langchain-core==1.2.12
langchain-google-genai==4.2.0
# langchain-huggingface==1.2.0 (Conflict)
langchain-openai==1.1.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.2
narwhals==1.39.1
networkx==3.4.2
numpy==2.4.2
openai==2.20.0
orjson==3.11.7
ormsgpack==1.12.2
osmnx==2.0.3
packaging==26.0
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
pipwin==0.5.2
playsound==1.2.2
plotly==6.1.0
prompt_toolkit==3.0.51
proto-plus==1.27.0
protobuf==5.29.6
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
py-ubjson==0.16.1
pyarrow==20.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.1
pyjsparser==2.7.1
PyJWT==2.12.1
pymongo==4.16.0
pyogrio==0.11.0
pyOpenSSL==26.0.0
pyparsing==3.2.3
pypiwin32==223
PyPrind==2.11.3
pyproj==3.7.1
pySmartDL==1.3.4
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pyttsx3==2.98
pytz==2025.2
pyvis==0.3.2
pywin32==308
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.2
rpds-py==0.25.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.15.3
service-identity==24.2.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.13.0
sqlparse==0.5.5
stack-data==0.6.3
streamlit==1.45.1
streamlit_folium==0.25.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tk==0.1.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==5.1.0
Twisted==25.5.0
txaio==25.12.2
typer==0.23.0
typer-slim==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.2
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.6.0
xxhash==3.6.0
xyzservices==2025.4.0
zope.interface==8.2
zstandard==0.25.0
supabase==2.13.0
